move pathList state and onChangePath function from <Path /> to <CreateEditPost />
